bot/requirements.txt
anyio==4.7.0
certifi==2024.12.14
dnspython==2.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pymongo==4.10.1
python-telegram-bot==21.9
sniffio==1.3.1
